requirements.txt
feedparser==6.0.12
google-generativeai==0.8.3
python-dotenv==1.1.1

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-cov==4.1.0

# Code quality (for CI/CD only - optional for local dev)
ruff==0.1.9
black==23.12.1
